pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biofake"
version = "0.1.0"
description = "Biomedical text deepfake robustness pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "biofake team" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.10",
  "PyYAML>=6.0",
  "scikit-learn>=1.6",
  "scipy>=1.12",
  "joblib>=1.4",
  "typer>=0.24",
  "streamlit>=1.45",
  "torch>=2.2",
  "transformers>=4.46",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[project.scripts]
biofake = "biofake.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

